% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vuk_palette.R
\name{get_vuk_palette}
\alias{get_vuk_palette}
\title{get_vuk_palette}
\usage{
get_vuk_palette(n)
}
\arguments{
\item{int}{n colors to return}
}
\value{
palette with n desired colors as named vector
}
\description{
Function to return color pallete using VUK visual identity, maximum is 8 colors
}
\examples{
get_vuk_palette()
}
\keyword{color}
\keyword{hex}
\keyword{pallete}
