% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vuk_palette.R
\name{show_vuk_palette}
\alias{show_vuk_palette}
\title{show_vuk_palette}
\usage{
show_vuk_palette(n)
}
\description{
Function to return color pallete using VUK visual identity, maximum is 8 colors
}
\examples{
show_vuk_palette()
}
\keyword{VUK}
\keyword{palette}
\keyword{show}
